cart renders from either local storage or state depending on user
