+++ python-project-starter-main/react-app/src/components/Shop.js
@@ -9,11 +9,6 @@ import ShopFilters from './ShopFilters';
 
 
 const Shop = () => {
-    let cartStorage = window.localStorage
-    let cart = cartStorage.getItem('cart')
-    if( cart === null){
-        cartStorage.setItem('cart', {});
-    }
     
     const history = useHistory()
     const dispatch = useDispatch()
