+++ python-project-starter-main/react-app/src/store/cart.js
@@ -25,6 +25,10 @@ const addOneToCart = (cart_item) => ({
 
 // =========== Thunk ========================
 export const getCartItems = (user_id) => async (dispatch) => {
+    if (!user_id) {
+        console.log('Warning: No user id provided to "getCartItems()"')
+        return
+    }
     const response = await fetch(`/api/users/${user_id}/cart`, {
         method: 'GET',
         headers: {
@@ -84,6 +88,7 @@ export default function reducer(state = initialState, action) {
         case ADD_TO_CART:
             let addedToExisting = false;
             newState.map((item) => {
+                console.log()
                 if(item.item_id === action.payload.item_id &&
                 item.item_color === action.payload.item_color && 
                 item.item_size === action.payload.item_size){
